Refactor model loading in evaluate_util.py and update execution counts in week1day4.ipynb
- Changed model loading to use model.cuda() instead of model.half().cuda() for better compatibility.
- Updated execution counts in week1day4.ipynb to reflect the correct order of cell execution.
- Adjusted output logs in week1day4.ipynb to match the new execution timestamps and messages.

+++ FIUBench/evaluate_util.py
@@ -659,7 +659,8 @@ def main(cfg):
             f"(e.g., LLaVA or Llama-3.2)."
         )
 
-    model.half().cuda()
+    # Move model to GPU
+    model.cuda()
 
     Path(cfg.save_dir).mkdir(parents=True, exist_ok=True)
